replace functions not directly related to doc.js inside helper
