+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,5 +1,4 @@
 ---
 name: Bug or improvement report
 about: Create a report to help us improve
-
 ---
