+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,5 +1,4 @@
 ---
 name: Feature request
 about: Suggest an idea for this project
-
 ---
